pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conversation-analytics"
version = "0.1.0"
description = "Advanced conversation detection and analysis system"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Zyra Audition", email = "contact@zyra-audition.com"}
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=0.24.2",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.2",
    "tqdm>=4.62.0",
    "psutil>=5.8.0",
    "scipy>=1.15.0",
    "PyYAML>=6.0.1",
    "openai==0.27.0",
    "anthropic==0.5.0",
    "ollama==0.1.0",
    "regex>=2024.11.6",
    "faiss-cpu==1.7.4",
    "spacy==3.5.0",
    "textblob>=0.17.1",
    "networkx==3.0.0",
    "hdbscan==0.8.29",
    "python-dateutil==2.8.2",
    "torch>=1.9.0",
    "transformers>=4.30.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.1",
    "bandit==1.7.5",
    "safety==2.3.0",
    "cryptography==41.0.0",
    "python-jose==3.3.0",
    "structlog==23.1.0",
    "sentry-sdk>=1.39.1",
    "prometheus-client==0.17.0",
    "python-json-logger>=2.0.7",
    "pyarrow>=14.0.1",
    "fastparquet>=2023.10.1",
    "cachetools>=5.3.2",
    "redis>=5.0.1",
] 